Add: MotionHandle.Time
