+++ src/LitMotion/Assets/LitMotion/Runtime/Internal/Error.cs
@@ -31,6 +31,11 @@ namespace LitMotion
             throw new ArgumentOutOfRangeException("Playback speed must be 0 or greater.");
         }
 
+        public static void TimeMustBeZeroOrGreater()
+        {
+            throw new ArgumentOutOfRangeException("Time must be 0 or greater.");
+        }
+
         public static void MotionNotExists()
         {
             throw new InvalidOperationException("Motion has been destroyed or no longer exists.");
