+++ src/LitMotion/Assets/LitMotion/Runtime/Internal/MotionHelper.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: b7b3337d974654c029ff5bfa7e15049c
